+++ apps/supaclk/metadata.json
@@ -1,0 +1,16 @@
+{ "id": "supaclk",
+  "name": "SUPACLOCK Pro ULTRA",
+  "version": "0.01",
+  "description": "SUPACLOCK Pro ULTRA, with four ClockInfo areas at the bottom. Tap them and swipe up/down and left/right to toggle between different information.",
+  "icon": "app.png",
+  "screenshots": [{"url":"screenshot.png"},{"url":"screenshot2.png"}],
+  "type": "clock",
+  "tags": "clock,clkinfo,clockinfo",
+  "supports" : ["BANGLEJS2"],
+  "readme": "README.md",
+  "dependencies" : { "clock_info":"module" },
+  "storage": [
+    {"name":"supaclk.app.js","url":"app.js"},
+    {"name":"supaclk.img","url":"app-icon.js","evaluate":true}
+  ]
+}
